requirements.txt
Authlib==1.5.2
blinker==1.9.0
click==8.1.8
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==0.19.0
requests==2.31.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
WTForms==3.1.1
tqdm==4.66.2
alembic==1.13.1
psycopg2-binary==2.9.1
email_validator==1.3.1
